+++ app/src/main/java/woowacourse/movie/domain/data/MovieData.kt
@@ -1,6 +1,6 @@
 package woowacourse.movie.domain.data
 
-import woowacourse.movie.domain.movieinfo.Movie
+import woowacourse.movie.domain.Movie
 
 interface MovieData {
     fun getAll(): List<Movie>
