--- app/src/main/java/woowacourse/movie/domain/movieinfo/RunningDate.kt
@@ -1,7 +1,0 @@
-package woowacourse.movie.domain.movieinfo
-
-import java.io.Serializable
-import java.time.LocalDate
-
-data class RunningDate(val startDate: LocalDate, val endDate: LocalDate) :
-    Serializable
