+++ app/src/main/java/woowacourse/movie/view/movielist/MovieListViewAdapter.kt
@@ -6,7 +6,7 @@ import android.view.View
 import android.view.ViewGroup
 import android.widget.BaseAdapter
 import woowacourse.movie.R
-import woowacourse.movie.domain.movieinfo.Movie
+import woowacourse.movie.domain.Movie
 import java.time.format.DateTimeFormatter
 
 class MovieListViewAdapter(
