+++ app/src/main/java/woowacourse/movie/view/viewmodel/MovieUIModel.kt
@@ -1,7 +1,7 @@
 package woowacourse.movie.view.viewmodel
 
 import androidx.annotation.DrawableRes
-import woowacourse.movie.domain.movieinfo.Movie
+import woowacourse.movie.domain.Movie
 import java.io.Serializable
 import java.time.LocalDate
